requirements.txt
PyQt5==5.15.7
Pillow==9.1.1
pyinstaller==5.1
appdirs==1.4.4
winshell==0.6
pywin32==304
